Fix Clyde erroneously culling on chunk rendering.
Fixes #912

+++ Robust.Client/Graphics/Clyde/Clyde.Rendering.cs
@@ -110,7 +110,7 @@ namespace Robust.Client.Graphics.Clyde
             // Calculate world-space AABB for camera, to cull off-screen things.
             var eye = _eyeManager.CurrentEye;
             var worldBounds = Box2.CenteredAround(eye.Position.Position,
-                _framebufferSize / EyeManager.PIXELSPERMETER * eye.Zoom);
+                _framebufferSize / (float)EyeManager.PIXELSPERMETER * eye.Zoom);
 
             using (DebugGroup("Lights"))
             {
